requirements.txt
blinker==1.9.0
cachelib==0.9.0
circuitbreaker==2.0.0
click==8.1.7
Deprecated==1.2.15
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Limiter==3.9.2
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mdurl==0.1.2
mysql-connector-python==9.1.0
ordered-set==4.1.0
packaging==24.2
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.0
